Ensure config file is not world-readable
Rather than relying on go-ini to create the config file, we'll manage the file
and make sure it has the permissions we want, so that we don't expose
credentials to other users on the system.

If the file exists, we chmod it first, and stop if chmod fails so nothing can
be exposed.

Add a test to confirm the permissions of a new file are what we expect,
and that if permissions get changed, rewriting the config file sets them back.