Add `--instance-role` flag to cluster up command
This allows the customer to specify the name of a custom IAM Role as the
instance role for
their EC2 Instance Profile. NOTE: If the correct IAM Policy
(`AmazonEC2ContainerServiceforEC2Role`) is not attached to the custom
instance
role, the create cluster command will succeed but no instances will be
launched.

If a custom role is specified, the `capability-iam` flag is not needed.

Closes #19.

```sh
$ make test
```
```sh
$ ecs-cli up --keypair my-keypair --size 2 --cluster my-test-cluster
--role ecsInstanceRole
```

```sh
$ ecs-cli up --keypair my-keypair --size 2 --cluster my-test-cluster
--instance-role foo --capability-iam
=> ERRO[0000] Error executing 'up': Cannot specify custom role when
'--capability-iam' flag is set

$ ecs-cli up --keypair my-keypair --size 2 --cluster my-test-cluster
=> ERRO[0000] Error executing 'up': You must either specify a custom
role with the '--instance-role' flag or set the '--capability-iam' flag
```