Add ecsParams onto context for ecsProject
Since the entity (i.e. task or service) does not take any arguments when
an action is called on it from the project (e.g. project.entity.up()),
it needs a way to access other parameters passed in through the
ecs-params file. Setting the ecsParams as a field on the project context
allows parameters such as networkConfiguration to be directly accessible by the
entity, since projectcontext is a field on the entity.