Implemented PS Command for Fargate Tasks and EC2 Tasks with Task Networking.
The Private IP for a task with Task Networking is obtained from the networkInterfaces section of the DescribeTask output. Fargate tasks can have public IPs, these are obtained by describing the ENI attached to the task. For both EC2 Tasks with Task Networking and Fargate tasks, the port mapping must be obtained by Describing the Task Definition that the Task uses.