[BUGFIX] Fix CFN template for creating Instance Profile
The previous conditional only created the EcsInstanceProfile in the
Cloudformation stack if the instance-role flag was blank and if the
launch-type was EC2 (CreateEcsInstanceRole condition). This caused a bug
when a custom instance-role was passed in with EC2 launch type, because
the Profile wouldn't be created, causing other resource creation to fail
with an "Unresolved resource dependencies" error due to
EcsInstanceProfile not being defined.

Fixes #394.