Fix YAML parse warning on networks field
Previously, the check for the "networks" field was based on the
NetworkConfigs attribute on the project to be empty. Libcompose changed
the behavior of the Parse method to always create a "default" network
config (See: https://github.com/docker/libcompose/commit/18dad7cb).

This caused a lot of spurious warnings for customers when they would try
to run a task even if the networks field was not specified in their
docker-compose files.

Fixes #406 and closes #237