Add force update flag
Values under 'run_params' in the ECSParams file are not considered
as part of the task definition and therefore will not trigger a new task
to be run if the the values under this field change. To ensure a better
developer experience, we are adding this flag to allow tasks/services to
be stopped and re-launched with the new run parameters.