[Bugfix] Allow container mem_limit to be null
Previously, if no mem_limt was set on a container, ecs-cli would default
it to 512mb, even if task size mem_limit was set. Additionally, if
mem_reservation only was set, mem_limit would be set to the same value.

This change leaves mem_limit unset (nil) in both cases.
Resolves #570, #606