Cleanup empty checks
These fields have the `yaml:",omitempty"` specification, so they will not be written
into the docker compose yaml as long as the values are empty.
For many of the ServiceConfig fields, an empty data structure is the
same as nil as far as the Marshaller is concerned, so explicitly
returning nil on an empty input is unnecessary.