fixed the issue where .env file is not used to substitute variables defined in V3 compose file
This commit fixes #726. The cause for the issue is the code path that handles variables defined in the `.env` file in Compose V3 is totally separated from the ones used by Compose V1, V2 (#839). So prior to this change, there was no code that actually handles reading from the `.env` file for Compose V3.