Include region in error message fields
This makes it easier to debug possible issues in which regions/profiles
may not be read from the same source. For example, if calls are made using an
AWS profile but the region is being read from the ecs-profile, i.e. if
there is a fallthrough in the order of resolution that is incompatible
for the requested resource.