Send optional fields as nil instead of "" to ECS
Previously when sending certain requests that contain optional fields to ECS through the Go AWS SDK, we always send empty strings instead of nil for those optional fields that are not set. The service treats such an action as an explicit intent and hence fails to plug in the default values from the service side.

This PR closes #927. Specifically, the repo 3.